requirements.txt
# =================================================================
# Core Data Science
# =================================================================
pandas
numpy
scikit-learn
torch


# =================================================================
# Utilities
# =================================================================
PyYAML
python-box
python-dotenv
ensure
tqdm


# =================================================================
# Google Cloud Platform (GCP) 
# =================================================================
google-cloud-storage

# =================================================================
# Experiment Tracking 
# =================================================================
mlflow
xgboost

# =================================================================
# Development & Visualization
# =================================================================
# notebook
# matplotlib
# seaborn

# =================================================================
# API Deployment (Optional for Stage 3)
# =================================================================
# Flask
# gunicorn

